# cap-sistema/docs/openapi/programas.yaml
openapi: 3.0.0
paths:
...
  /v1/programas/embarazo:
    get:
      operationId: EmbarazoController_listar
      parameters:
        - name: estado
          required: false
          in: query
          schema:
            enum:
              - ACTIVO
              - EGRESADO
              - ABANDONO
              - FALLECIDO
              - TRASLADADO
            type: string
        - name: riesgo
          required: false
          in: query
          schema:
            enum:
              - BAJO
              - ALTO
            type: string
        - name: comunidadId
          required: false
          in: query
          schema:
            format: uuid
            type: string
        - name: pagina
          required: false
          in: query
          description: Empieza en 1. Por defecto 1.
          schema:
            example: 1
            type: number
        - name: tamano
          required: false
          in: query
          description: Por defecto 25. El servidor recorta cualquier valor mayor a 100.
          schema:
            example: 25
            type: number
      responses:
        "200":
          description: Listado paginado.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginaDto"
                  - type: object
                    required:
                      - datos
                    properties:
                      datos:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProgramaEmbarazoResumenDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: &a4
        - bearer: []
      summary: Seguimientos, ordenados por fecha probable de parto
      tags: &a5
        - programa-embarazo
    post:
      description: La FPP y la clasificacion de riesgo las calcula el sistema a partir
        de la FUM.
      operationId: EmbarazoController_inscribir
      parameters:
        - name: authorization
          required: true
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InscribirEmbarazoDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbarazoInscritoDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a4
      summary: Inscribe un embarazo
      tags: *a5
  /v1/programas/embarazo/alto-riesgo:
    get:
      operationId: EmbarazoController_altoRiesgo
      parameters:
        - name: pagina
          required: false
          in: query
          description: Empieza en 1. Por defecto 1.
          schema:
            example: 1
            type: number
        - name: tamano
          required: false
          in: query
          description: Por defecto 25. El servidor recorta cualquier valor mayor a 100.
          schema:
            example: 25
            type: number
      responses:
        "200":
          description: Listado paginado.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginaDto"
                  - type: object
                    required:
                      - datos
                    properties:
                      datos:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProgramaEmbarazoResumenDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a4
      summary: Embarazos activos clasificados como de alto riesgo
      tags: *a5
  /v1/programas/embarazo/{id}:
    get:
      operationId: EmbarazoController_obtener
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramaEmbarazoDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "404":
          description: El recurso solicitado no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a4
      summary: Seguimiento por su identificador
      tags: *a5
  /v1/programas/embarazo/{id}/controles:
    get:
      operationId: EmbarazoController_listarControles
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: pagina
          required: false
          in: query
          description: Empieza en 1. Por defecto 1.
          schema:
            example: 1
            type: number
        - name: tamano
          required: false
          in: query
          description: Por defecto 25. El servidor recorta cualquier valor mayor a 100.
          schema:
            example: 25
            type: number
      responses:
        "200":
          description: Listado paginado.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginaDto"
                  - type: object
                    required:
                      - datos
                    properties:
                      datos:
                        type: array
                        items:
                          $ref: "#/components/schemas/ControlPrenatalDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "404":
          description: El recurso solicitado no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a4
      summary: Controles prenatales, lo mas reciente primero
      tags: *a5
    post:
      description: Las semanas de gestacion, las senales de alarma y la fecha del
        proximo control las calcula el sistema.
      operationId: EmbarazoController_registrarControl
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrarControlPrenatalDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlPrenatalDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "404":
          description: El recurso solicitado no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a4
      summary: Registra un control prenatal
      tags: *a5
  /v1/programas/embarazo/{id}/cierre:
    patch:
      operationId: EmbarazoController_cerrar
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: "Sin semanasGestacion: el seguimiento ya cerro, la cuenta dejo de
            correr."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramaEmbarazoBaseDto"
        "400":
          description: La informacion enviada no es valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "401":
          description: Falta el token, expiro o no es valido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "403":
          description: El rol de la cuenta no tiene permiso sobre este recurso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "404":
          description: El recurso solicitado no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
        "500":
          description: Error inesperado. El mensaje real queda en los logs, no se expone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespuestaErrorDto"
      security: *a4
      summary: Cierra el seguimiento con su resultado
      tags: *a5
info:
  title: Servicio programas - CAP Purulha
  description: >-
    Contrato del servicio programas.


    Generado desde el codigo. No editar a mano: los cambios se pierden en la
    siguiente exportacion. Fuente de verdad: arquitectura-cap-purulha.md
  version: "1.0"
  contact: {}
...
servers:
  - url: http://localhost:3003
    description: Desarrollo local
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    PaginaDto:
      type: object
      properties:
        pagina:
          type: number
          description: Pagina devuelta, empezando en 1.
          example: 1
        tamano:
          type: number
          description: Registros por pagina. El servidor recorta cualquier valor mayor al
            maximo.
          maximum: 100
          example: 25
        total:
          type: number
          description: Total de registros que cumplen el filtro.
          example: 137
        totalPaginas:
          type: number
          description: Total de paginas. Nunca es menor que 1.
          example: 6
      required:
        - pagina
        - tamano
        - total
        - totalPaginas
...
    UltimoControlPrenatalDto:
      type: object
      properties:
        fecha:
          format: date-time
          type: string
        semanasGestacion:
          type: number
          example: 24
        proximoControl:
          type: string
          format: date-time
          nullable: true
        alertas:
          description: Senales de alarma detectadas por el sistema.
          type: array
          items:
            type: string
      required:
        - fecha
        - semanasGestacion
        - proximoControl
        - alertas
    ProgramaEmbarazoResumenDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pacienteId:
          type: string
          format: uuid
        comunidadId:
          type: string
          format: uuid
        fum:
          format: date-time
          type: string
        fpp:
          format: date-time
          type: string
        riesgo:
          type: string
          enum:
            - BAJO
            - ALTO
        motivoRiesgo:
          type: string
          nullable: true
        estado:
          type: string
          enum: &a8
            - ACTIVO
            - EGRESADO
            - ABANDONO
            - FALLECIDO
            - TRASLADADO
        numeroGestacion:
          type: number
          example: 3
        semanasGestacion:
          type: number
          example: 24
        ultimoControl:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/UltimoControlPrenatalDto"
      required:
        - id
        - pacienteId
        - comunidadId
        - fum
        - fpp
        - riesgo
        - motivoRiesgo
        - estado
        - numeroGestacion
        - semanasGestacion
        - ultimoControl
    ProgramaEmbarazoDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pacienteId:
          type: string
          format: uuid
        comunidadId:
          type: string
          format: uuid
        fum:
          format: date-time
          type: string
          description: Fecha de ultima menstruacion.
        fpp:
          format: date-time
          type: string
          description: "Fecha probable de parto: FUM + 280 dias (regla de Naegele)."
        numeroGestacion:
          type: number
          example: 3
        partosPrevios:
          type: number
          example: 2
        riesgo:
          type: string
          enum: &a9
            - BAJO
            - ALTO
        motivoRiesgo:
          type: string
          nullable: true
          maxLength: 300
          description: Por que se clasifico como alto riesgo, para que el siguiente turno
            sepa a que atenerse.
        estado:
          type: string
          enum: *a8
        resultado:
          type: string
          enum: &a10
            - PARTO_NORMAL
            - CESAREA
            - ABORTO
            - OBITO
            - TRASLADO
            - OTRO
          nullable: true
        fechaCierre:
          type: string
          format: date-time
          nullable: true
        inscritoPor:
          type: string
          format: uuid
        creadoEn:
          format: date-time
          type: string
        semanasGestacion:
          type: number
          description: Calculada al responder sobre el dia local, no almacenada.
          example: 24
      required:
        - id
        - pacienteId
        - comunidadId
        - fum
        - fpp
        - numeroGestacion
        - partosPrevios
        - riesgo
        - motivoRiesgo
        - estado
        - resultado
        - fechaCierre
        - inscritoPor
        - creadoEn
        - semanasGestacion
    InscribirEmbarazoDto:
      type: object
      properties:
        pacienteId:
          type: string
        fum:
          format: date-time
          type: string
          example: 2026-01-15
          description: Fecha de ultima menstruacion. De aqui salen la FPP y las semanas de
            gestacion.
        numeroGestacion:
          type: number
          default: 1
          minimum: 1
        partosPrevios:
          type: number
          default: 0
      required:
        - pacienteId
        - fum
    EmbarazoInscritoDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pacienteId:
          type: string
          format: uuid
        comunidadId:
          type: string
          format: uuid
        fum:
          format: date-time
          type: string
          description: Fecha de ultima menstruacion.
        fpp:
          format: date-time
          type: string
          description: "Fecha probable de parto: FUM + 280 dias (regla de Naegele)."
        numeroGestacion:
          type: number
          example: 3
        partosPrevios:
          type: number
          example: 2
        riesgo:
          type: string
          enum: *a9
        motivoRiesgo:
          type: string
          nullable: true
          maxLength: 300
          description: Por que se clasifico como alto riesgo, para que el siguiente turno
            sepa a que atenerse.
        estado:
          type: string
          enum: *a8
        resultado:
          type: string
          enum: *a10
          nullable: true
        fechaCierre:
          type: string
          format: date-time
          nullable: true
        inscritoPor:
          type: string
          format: uuid
        creadoEn:
          format: date-time
          type: string
        semanasGestacion:
          type: number
          description: Calculada al responder sobre el dia local, no almacenada.
          example: 24
        motivosRiesgo:
          description: Vacio cuando el riesgo es bajo.
          example:
            - Edad menor de 18 anios
          type: array
          items:
            type: string
      required:
        - id
        - pacienteId
        - comunidadId
        - fum
        - fpp
        - numeroGestacion
        - partosPrevios
        - riesgo
        - motivoRiesgo
        - estado
        - resultado
        - fechaCierre
        - inscritoPor
        - creadoEn
        - semanasGestacion
        - motivosRiesgo
    ControlPrenatalDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        programaId:
          type: string
          format: uuid
        fecha:
          format: date-time
          type: string
        semanasGestacion:
          type: number
          description: Calculadas a partir de la FUM y la fecha del control.
          example: 24
        pesoKg:
          type: string
          format: decimal
          nullable: true
          example: "58.30"
        sistolica:
          type: number
          nullable: true
        diastolica:
          type: number
          nullable: true
        alturaUterinaCm:
          type: string
          format: decimal
          nullable: true
          description: Altura uterina en centimetros.
          example: "24.0"
        fcf:
          type: number
          nullable: true
          description: Frecuencia cardiaca fetal, latidos por minuto.
          example: 140
        edema:
          type: boolean
          nullable: true
        alertas:
          description: Senales de alarma detectadas por el sistema en este control.
          type: array
          items:
            type: string
        proximoControl:
          type: string
          format: date-time
          nullable: true
        registradoPor:
          type: string
          format: uuid
        observaciones:
          type: string
          nullable: true
          description: Descifradas.
      required:
        - id
        - programaId
        - fecha
        - semanasGestacion
        - pesoKg
        - sistolica
        - diastolica
        - alturaUterinaCm
        - fcf
        - edema
        - alertas
        - proximoControl
        - registradoPor
        - observaciones
    RegistrarControlPrenatalDto:
      type: object
      properties:
        pesoKg:
          type: number
          minimum: 30
          maximum: 200
        sistolica:
          type: number
          minimum: 50
          maximum: 300
        diastolica:
          type: number
          minimum: 30
          maximum: 200
        alturaUterinaCm:
          type: number
          minimum: 5
          maximum: 50
          description: Altura uterina en centimetros
        fcf:
          type: number
          minimum: 60
          maximum: 220
          description: Frecuencia cardiaca fetal
        edema:
          type: boolean
        observaciones:
          type: string
        fecha:
          format: date-time
          type: string
    ProgramaEmbarazoBaseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pacienteId:
          type: string
          format: uuid
        comunidadId:
          type: string
          format: uuid
        fum:
          format: date-time
          type: string
          description: Fecha de ultima menstruacion.
        fpp:
          format: date-time
          type: string
          description: "Fecha probable de parto: FUM + 280 dias (regla de Naegele)."
        numeroGestacion:
          type: number
          example: 3
        partosPrevios:
          type: number
          example: 2
        riesgo:
          type: string
          enum: *a9
        motivoRiesgo:
          type: string
          nullable: true
          maxLength: 300
          description: Por que se clasifico como alto riesgo, para que el siguiente turno
            sepa a que atenerse.
        estado:
          type: string
          enum: *a8
        resultado:
          type: string
          enum: *a10
          nullable: true
        fechaCierre:
          type: string
          format: date-time
          nullable: true
        inscritoPor:
          type: string
          format: uuid
        creadoEn:
          format: date-time
          type: string
      required:
        - id
        - pacienteId
        - comunidadId
        - fum
        - fpp
        - numeroGestacion
        - partosPrevios
        - riesgo
        - motivoRiesgo
        - estado
        - resultado
        - fechaCierre
        - inscritoPor
        - creadoEn
    RespuestaErrorDto:
      type: object
      properties:
        codigo:
          type: string
          description: Codigo estable. El cliente decide que hacer segun este valor, no
            segun el mensaje.
          enum:
            - VALIDACION
            - NO_AUTENTICADO
            - SIN_PERMISO
            - NO_ENCONTRADO
            - CONFLICTO
            - DEMASIADAS_PETICIONES
            - ERROR_INTERNO
          example: VALIDACION
        mensaje:
          type: string
          description: Mensaje en espanol, apto para mostrarse tal cual al usuario.
          example: La informacion enviada no es valida.
        detalles:
          description: Detalle por campo. Presente sobre todo en errores de validacion.
          example:
            - dpi debe tener 13 digitos
          type: array
          items:
            type: string
        trazaId:
          type: string
          description: Identificador de correlacion. Es lo que se busca en los logs para
            esta peticion.
          example: 7f3a1c2e-9b4d-4e21-8a55-0c1d2e3f4a5b
        ruta:
          type: string
          example: /v1/pacientes
        fecha:
          type: string
          format: date-time
          example: 2026-08-26T15:04:05.000Z
      required:
        - codigo
        - mensaje
        - trazaId
        - ruta
        - fecha